pipeline/requirements.txt
ffmpeg-python
pandas==2.2.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tokenizers<0.16  # Ensure compatibility with whisperx
pytubefix
whisperx==3.1.5
pyannote.audio
google-api-python-client
